required cast to long-enough unsigned
git-svn-id: svn+ssh://svn.cprover.org/var/lib/cvs/satabs/trunk@241 c82bf971-7879-4458-aa57-88e1f0fec4b8

+++ src/satabs/refiner/transition_cache.h
@@ -49,7 +49,7 @@ class transition_cachet
     {
       size_t operator()(const entryt &e) const
       {
-        return ((unsigned long)&(*e.pc))^e.from.size()^e.to.size()
+        return ((unsigned long long)&(*e.pc))^e.from.size()^e.to.size()
           ^e.from_passive.size()^e.to_passive.size();
       }
     };
